refactor(actions): rename _item_helpers to item_helpers
